backend/requirements-dev.txt
# ============================================
# NOUFAL Engineering System - Development Dependencies
# ============================================
# تاريخ الإنشاء: 2025-11-06
# للاستخدام أثناء التطوير فقط
# ============================================

# Include all production dependencies
-r requirements-full.txt

# Development Tools
# =====================================
ipython==8.19.0                 # Enhanced Python Shell
jupyter==1.0.0                  # Jupyter Notebooks
notebook==7.0.6                 # Jupyter Notebook Server

# Debugging
# =====================================
ipdb==0.13.13                   # IPython Debugger
pdb-attach==3.1.0               # Remote Debugging
flask-debugtoolbar==0.14.1      # Flask Debug Toolbar

# Testing Enhancements
# =====================================
pytest-watch==4.2.0             # Auto-run Tests
pytest-xdist==3.5.0             # Parallel Testing
coverage==7.4.0                 # Code Coverage

# Documentation
# =====================================
sphinx==7.2.6                   # Documentation Generator
sphinx-rtd-theme==2.0.0         # ReadTheDocs Theme

# Database Tools
# =====================================
pgcli==4.0.1                    # PostgreSQL CLI
mycli==1.27.2                   # MySQL CLI (optional)

# API Testing
# =====================================
httpie==3.2.2                   # HTTP Client CLI
postman-cli==1.0.0              # Postman CLI (optional)

# Performance Profiling
# =====================================
py-spy==0.3.14                  # Sampling Profiler
memory-profiler==0.61.0         # Memory Usage

# Code Generation
# =====================================
cookiecutter==2.5.0             # Project Templates

# ============================================
